requirements.txt
nibabel>=5.1.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
scipy>=1.11.0
opencv-python>=4.8.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.66.0
monai>=1.2.0
PyQt5>=5.15.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0+cu121
torchvision==0.16.0+cu121
torchaudio==2.1.0+cu121 